Support another weird Swiss DL MRZ
This time I found a version with an extra character in the version
field (presumably), which makes the long lines 31 characters per line.

Here are some examples for this (as long as the link works):
https://www.astra.admin.ch/dam/astra/de/dokumente/abteilung_strassenverkehrallgemein/information-schweizer-fuehrerausweis.pdf.download.pdf/Information%20Schweizer%20F%C3%BChrerausweis.pdf

Curiously, the last example on the last page is again only 30 characters
per line. I don't know what kind of cheese they eat in Switzerland, but
they should definitely stop eating it.